LayoutTests/platform/chromium/inspector/console/command-line-api-getEventListeners-expected.txt

Tests getEventListeners() method of console command line API.

 
- inner -
keydown: {
    0: {
        listener: function listener1() { }
        useCapture: false
    }
    1: {
        listener: function listener2() { }
        useCapture: true
    }
}
- outer -
mousemove: {
    0: {
        listener: function listener1() { }
        useCapture: false
    }
}
keydown: {
    0: {
        listener: function listener2() { }
        useCapture: true
    }
}
mousedown: {
    0: {
        listener: function listener2() { }
        useCapture: true
    }
}
- attribute event listeners -
mouseover: {
    0: {
        listener: function onmouseover(event) { listener2() }
        useCapture: false
    }
}
click: {
    0: {
        listener: function onclick(event) { alert(1) }
        useCapture: false
    }
}
- empty -
- invalid -
- object -
undefined
- null -
undefined
- undefined -
undefined
- window -
undefined
